Django-Middleware-0x03/requirements.txt
amqp==5.3.1
asgiref==3.8.1
Babel==2.15.0
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
et-xmlfile==1.1.0
inflection==0.5.1
kombu==5.5.3
mysql-connector-python==9.3.0
openpyxl==3.1.5
packaging==25.0
plotly==5.16.1
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tenacity==8.2.3
tkcalendar==1.6.1
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
